StartBusiness/requirements.txt
amqp==5.2.0
asgiref==3.7.2
BareNecessities==0.2.8
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.7.0
boto3==1.34.40
botocore==1.34.40
bs4==0.0.1
cachetools==5.3.2
celery==5.4.0
certifi==2023.7.22
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.7.1
distlib==0.3.7
<<<<<<< HEAD

=======
>>>>>>> e1b890517efb34e16626715115802d3945e9f84a
django-cors-headers==4.3.0
django-filter==23.3
django-phone-field==1.8.1
django-rest-swagger==2.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
<<<<<<< HEAD

=======
>>>>>>> e1b890517efb34e16626715115802d3945e9f84a
dnspython==2.4.2
drf-yasg==1.21.7
ecdsa==0.18.0
filelock==3.12.4
Flask==3.0.0
google-api-core==2.18.0
google-api-python-client==2.105.0
google-auth==2.29.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.61.0
gunicorn==22.0.0
h11==0.14.0
httplib2==0.22.0
icalendar==5.0.10
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jose==1.0.0
kombu==5.3.7
Mail==2.1.0
MarkupSafe==2.1.3
mongoengine==0.27.0
oauth2client==4.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.24.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymongo==3.12.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-nmap==0.7.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.1
setuptools==68.2.2
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
soupsieve==2.5
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.29.0
vine==5.1.0
virtualenv==20.24.5
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
Werkzeug==3.0.1
whitenoise==6.6.0
